feat(Templates): metaBody for sections and formats
