+++ src/templates/Article/docs.md
@@ -24,6 +24,7 @@ const schema = createArticleSchema({
   - `href` String, target url or path
   - `passHref` Boolean, indicates this will eventually end in an a tag and you may overwrite href
 - `getVideoPlayerProps`, a [prop getter](https://blog.kentcdodds.com/how-to-give-rendering-control-to-users-with-prop-getters-549eaef76acf) for the video player. Make sure to forward, modified or unmodified, the props that are passed to the function.
+- `metaBody`, use the meta font for body text
 
 # Example
