+++ src/templates/Section/docs.md
@@ -14,6 +14,7 @@ Defaults:
 - `customMetaFields`, always adds repo refs for `discussion`, `dossier` and `format` settings with a `kind` (font) and `color` field.
 - `series`, false
 - `darkMode`, false
+- `metaBody`, true
 
 # Example
