+++ src/templates/Section/index.js
@@ -24,6 +24,7 @@ const createSchema = ({
   titleMargin = true,
   titleBlockRule,
   getPath = ({ slug }) => `/${(slug || '').split('/').pop()}`,
+  metaBody = true,
   ...args
 } = {}) => {
   return createArticleSchema({
@@ -55,6 +56,7 @@ const createSchema = ({
     series,
     darkMode,
     paynotes,
+    metaBody,
     titleBlockRule: titleBlockRule || {
       matchMdast: matchZone('TITLE'),
       component: ({ children, ...props }) => (
